lightsaber-backend/requirements.txt
Werkzeug
requests
Flask
face_recognition
Pillow
pymongo
flask-cors
python-dotenv
jsonify